Fix merging of objects that straddle tile boundaries
I was ignoring the fact that objects will always end up
getting cut on a tile boundary, and this will cause their
IoU with their neighbours to be very small, particularly
if the NN clips the objects bounds to its input window.

Now, we account for this, and merge objects across
tiles with a much better metric.